[Claude Code] Refactor state data model.
Prompt:

> I've made some changes to the QuizState type in
> src/quizStateReducer.ts, which I believe result in a more
> coherent and easier-to-understand state schema. You can
> use `git diff` to see what they are.
>
> Can you modify the rest of the code to use this new
> state schema?

This had mixed results. Possibly because I made multiple different changes
to the data model, rather than one at a time, Claude refactored the code to
always pick the very first database row for initial state (since it can no
longer be undefined). When I pointed this out, it didn't run tests afterwards,
so it didn't notice that it broke the app by calling a function before it was
used.

This... really feels like having an intern. Except an intern would grow
and change as I educated them, and this doesn't (yet?).