Decompose jamos into keystrokes.
When nearing the end of typing "김민지", the IME has composed
"김믽", and because the last jamo doesn't match any of the jamos
in the answer, the app doesn't count it as correct.

This PR decomposes all the jamos into their constituent keystrokes,
instead of jamos, which fixes the bug.

I will rename the files in my next commit; I didn't want to rename
them here because it might impair the diff.